Pull mouse mode handling into its own module
